requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-core
openai
python-dotenv
line-bot-sdk
pydantic
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2